pyproject.toml
[project]
name = "docchat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.5",
    "docling>=2.82.0",
    "groq>=0.37.1",
    "huggingface-hub[hf-xet]>=0.36.2",
    "langchain>=1.2.13",
    "langchain-chroma>=1.1.0",
    "langchain-classic>=1.0.3",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.22",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.1",
    "langchain-openai>=1.1.12",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.3",
    "loguru>=0.7.3",
    "openai>=2.30.0",
    "pydantic-settings>=2.13.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "rank-bm25>=0.2.2",
    "sentence-transformers>=5.3.0",
    "streamlit>=1.55.0",
    "torch>=2.11.0",
    "transformers>=4.57.6",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"
